+++ 3o14/components/common/ContentRenderer/types.ts
@@ -1,0 +1,9 @@
+import type { MixedStyleDeclaration } from 'react-native-render-html';
+
+export interface ContentRendererProps {
+  content: string;
+  width: number;
+  tagsStyles: Record<string, MixedStyleDeclaration>;
+  renderersProps: any;
+  systemFonts: string[];
+}
